CI Attempt #6 - Shift helper functions to package
