+++ Telegram Bot/Patch Notes/bot_featuresToAdd.txt
