+++ Telegram Bot/Patch Notes/bot_glitchesToFix.txt
